Moved create_path function into a separate package for utility functions.
