+++ db/dump.py
@@ -11,13 +11,13 @@ from __future__ import print_function
 from collections import defaultdict
 from datetime import datetime
 from db import create_cursor, commit
+import utils.path
 import db
 import subprocess
 import tempfile
 import logging
 import tarfile
 import shutil
-import errno
 import os
 
 DUMP_CHUNK_SIZE = 1000
@@ -74,7 +74,7 @@ def dump_db(location, threads=None, incremental=False, dump_id=None):
     Returns:
         Path to created dump.
     """
-    create_path(location)
+    utils.path.create_path(location)
     time_now = datetime.today()
 
     if incremental:
@@ -112,7 +112,7 @@ def dump_db(location, threads=None, incremental=False, dump_id=None):
                     arcname=os.path.join(archive_name, "COPYING"))
 
             archive_tables_dir = os.path.join(temp_dir, "abdump", "abdump")
-            create_path(archive_tables_dir)
+            utils.path.create_path(archive_tables_dir)
             _copy_tables(archive_tables_dir, start_t, end_t)
             tar.add(archive_tables_dir, arcname=os.path.join(archive_name, "abdump"))
 
@@ -225,7 +225,7 @@ def dump_lowlevel_json(location, incremental=False, dump_id=None):
     Returns:
         Path to created low level JSON dump.
     """
-    create_path(location)
+    utils.path.create_path(location)
 
     if incremental:
         dump_id, start_time, end_time = prepare_incremental_dump(dump_id)
@@ -321,7 +321,7 @@ def dump_highlevel_json(location, incremental=False, dump_id=None):
     Returns:
         Path to created high-level JSON dump.
     """
-    create_path(location)
+    utils.path.create_path(location)
 
     if incremental:
         dump_id, start_time, end_time = prepare_incremental_dump(dump_id)
@@ -476,16 +476,6 @@ def _get_incremental_dump_timestamp(dump_id=None):
             cursor.execute("SELECT created FROM incremental_dumps ORDER BY id DESC")
         row = cursor.fetchone()
     return row[0] if row else None
-
-
-def create_path(path):
-    """Creates a directory structure if it doesn't exist yet."""
-    try:
-        os.makedirs(path)
-    except OSError as exception:
-        if exception.errno != errno.EEXIST:
-            raise Exception("Failed to create directory structure %s. Error: %s" %
-                            (path, exception))
 
 
 class NoNewData(Exception):
